requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
python-dotenv
pydantic
duckduckgo-search
wikipedia
pandas
numpy
scikit-learn
nltk
spacy
pypdf
faiss-cpu
sentence-transformers
chromadb
tiktoken
regex
openai
unstructured
tabulate
beautifulsoup4
tqdm
backoff
memory-profiler